Added reference counter to tasks
The reference counter is implemented to avoid that tasks are free'd
while there are streams that want to set an event flag in the task struct.
The ref count is an atomic variable set to 1 upon creation, incremented
each time a stream is opened, and decremented upon TaskDestroy.
TaskDestroy is called from the worker thread as well as from StreamDestroy,
if the consumer/producer is set.